# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tianquan 天权 API",
    "description": "Permission-aware enterprise knowledge agent.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/revoke": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke",
        "operationId": "revoke_admin_revoke_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ACLChangeModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/grant": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Grant",
        "operationId": "grant_admin_grant_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ACLChangeModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ACLChangeModel": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "previous_version": {
            "type": "integer",
            "title": "Previous Version"
          },
          "new_version": {
            "type": "integer",
            "title": "New Version"
          },
          "version_transition": {
            "type": "string",
            "title": "Version Transition"
          },
          "change": {
            "type": "string",
            "title": "Change"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "subject_kind": {
            "type": "string",
            "title": "Subject Kind"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "previous_version",
          "new_version",
          "version_transition",
          "change",
          "subject",
          "subject_kind"
        ],
        "title": "ACLChangeModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevokeGrantRequest": {
        "properties": {
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "A user_id or a role name."
          },
          "subject_kind": {
            "type": "string",
            "pattern": "^(user|role)$",
            "title": "Subject Kind",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "resource_id",
          "subject"
        ],
        "title": "RevokeGrantRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}